move documentation folder to modules
