+++ package.json
@@ -12,7 +12,7 @@
     "build:backend": "tsc -p src/apps/backend/tsconfig.json --outDir dist",
     "build:frontend": "webpack --output-path dist/public --config src/apps/frontend/webpack.prod.js",
     "fmt": "prettier --write \"**/*.{js,ts,tsx}\" --ignore-path .prettierignore",
-    "generate-documentation": "ts-node src/apps/backend/documentation/documentation-script.ts",
+    "generate-documentation": "ts-node src/apps/backend/modules/documentation/documentation-script.ts",
     "test": "cross-env NODE_ENV=production NODE_CONFIG_ENV=testing mocha --exit -r ts-node/register --project tsconfig.mocha.json test/**/*.spec.ts",
     "test:e2e": "cross-env NODE_ENV=production NODE_CONFIG_ENV=e2e concurrently --success first --kill-others --kill-others-on-fail \"wait-on tcp:8080 && cypress run\" \"npm start\"",
     "lint": "eslint .",
