+++ src/apps/backend/modules/documentation/internals/types.ts
